Refactor World
